badgerdb-sampler: Split code into consensus and runtime analysis
